OPENBSD_4_0

+++ dev/pcmcia/com_pcmcia.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: com_pcmcia.c,v 1.44 2005/12/13 23:07:31 mickey Exp $	*/
+/*	$OpenBSD: com_pcmcia.c,v 1.46 2006/05/28 16:49:46 fkr Exp $	*/
 /*	$NetBSD: com_pcmcia.c,v 1.15 1998/08/22 17:47:58 msaitoh Exp $	*/
 
 /*
@@ -124,7 +124,6 @@
 
 #include <dev/ic/comreg.h>
 #if NPCCOM > 0
-#include <dev/isa/isavar.h>
 #include <i386/isa/pccomvar.h>
 #endif
 #if NCOM > 0
@@ -318,7 +317,8 @@ found:
 		return;
 	}
 
-	printf(" port 0x%lx/%d", psc->sc_pcioh.addr, psc->sc_pcioh.size);
+	printf(" port 0x%lx/%lu", psc->sc_pcioh.addr,
+	    (u_long)psc->sc_pcioh.size);
 
 	sc->sc_iobase = -1;
 	sc->enable = com_pcmcia_enable;
